requirements.txt
pillow
opencv-contrib-python
open3d
